file restructing and cleaning
