+++ did_ethr_linked_resources/src/registar.rs
@@ -35,7 +35,7 @@ where
         let resource = self
             .registry
             .create_or_update_resource(self.signer.clone(), &did, resource_input)
-            .await;
+            .await?;
 
         let metadata_node = self
             .resolver
