+++ eth-anoncreds-rust-demo/src/anoncreds_method/mod.rs
@@ -1,4 +1,3 @@
-pub mod ethers_client;
 pub mod ledger_data_transformer;
 pub mod registar;
 pub mod resolver;
