+++ eth-anoncreds-rust-demo/src/roles/mod.rs
@@ -1,0 +1,3 @@
+pub mod holder;
+pub mod issuer;
+pub mod verifier;
